% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{get_resource}
\alias{get_resource}
\title{Get a resource from data.gov.au}
\usage{
get_resource(id = NULL, url = NULL, return.type = NULL, ...)
}
\arguments{
\item{id, url}{The id from \url{https://data.gov.au} specifying the resource,
or the URL for the direct download. Either the \code{id} or and only or the \code{url} must be given; if both
are provided, the url is used, with a warning.}

\item{return.type}{In what format should the data be returned.
Currently, only \code{NULL} is supported, which returns the data
in whatever form the underlying function for the particular resource
format returns. (For example, \code{csv}'s are returned as \code{data.table}s;
\code{xlsx} files are returned as \code{tibbles}.)}

\item{...}{Passed to the function which reads the resource, once downloaded.}
}
\description{
A function mainly for internal use, that retrieves the resource
(such as a table of data) and returning an object for analysis in the R
session (as opposed to downloading the file for later access)
}
