% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagovau-package.R
\docType{package}
\name{package-data.gov.au}
\alias{datagovau}
\alias{datagovau-package}
\alias{package-data.gov.au-package}
\title{Retrieve data from data.gov.au}
\description{
Provides an interface for R users to search and load data
from the Australian Government's data portal, <https://data.gov.au>.
}
\author{
\strong{Maintainer}: Jonathan Carroll \email{rpkg@jcarroll.com.au} (@jonocarroll)

Authors:
\itemize{
  \item Maria Dmitrijeva (@marianess)
  \item Jeffrey O. Hanson (@jeffreyhanson)
  \item Rob J. Hyndman (@robjhyndman)
  \item Simon Lyons (@SimonLyons)
  \item Cameron Roach (@camroach87)
  \item Adam H. Sparks (@adamhsparks)
  \item Matt Sutton (@matt-sutton)
  \item Peter Ellis (@ellis2013nz)
}

Other contributors:
\itemize{
  \item Jacinta Holloway [contributor]
}

}
\keyword{package}
