% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understanding_metadata.R
\name{search_data}
\alias{search_data}
\title{Search the data.gov.au metadata}
\usage{
search_data(query = "name:location", limit = 10)
}
\arguments{
\item{query}{character string of field to be searched and string to search it for, separated by a colon.}

\item{limit}{number of rows of metadata to return (each row represents a single package of datasets).}
}
\value{
A tibble of metadata with 30 columns
}
\description{
Search the data.gov.au metadata
}
\details{
Good fields to search include name, description, format, url, licence_id
}
\examples{

\dontrun{
if (requireNamespace("dplyr", quietly = TRUE)) {
res <- search_data("name:water", limit = 20)
water_data <- res \%>\% filter(can_use == "yes") \%>\% slice(2) \%>\% get_data
head(water_data[[1]])
}
}
}
\author{
Jonathan Carroll
}
